chore: Move public header to Includes/
